Painovoima toimii

+++ Avaruustomuttaja/src/main/java/avaruustomuttaja/kappale/Kappale.java
@@ -5,18 +5,18 @@ import java.awt.Graphics;
 public class Kappale {
 
     private double massa;
-    private int paikkax;
-    private int paikkay;
+    private double paikkax;
+    private double paikkay;
     private Vektori nopeus;
 
-    public Kappale(double massa, int paikkax, int paikkay) {
+    public Kappale(double massa, double paikkax, double paikkay) {
         this.massa = massa;
         this.paikkax = paikkax;
         this.paikkay = paikkay;
         this.nopeus = new Vektori(0, 0);
     }
 
-    public Kappale(double massa, int paikkax, int paikkay, double nopeusx, double nopeusy) {
+    public Kappale(double massa, double paikkax, double paikkay, double nopeusx, double nopeusy) {
         this.massa = massa;
         this.paikkax = paikkax;
         this.paikkay = paikkay;
@@ -31,11 +31,11 @@ public class Kappale {
         this.massa = massa;
     }
 
-    public int getPaikkaX() {
+    public double getPaikkaX() {
         return paikkax;
     }
 
-    public int getPaikkaY() {
+    public double getPaikkaY() {
         return paikkay;
     }
